Corrigindo testes e adicionando pacote mongo de testes
