+++ .gitignore
@@ -1,2 +1,3 @@
 server/node_modules/
 .env-prod
+server/coverage/
