+++ server/globalConfig.json
@@ -1,0 +1,1 @@
+{"mongoUri":"mongodb://127.0.0.1:38847/jest?","mongoDBName":"jest"}
