+++ server/jest-mongodb-config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    mongodbMemoryServerOptions: {
+        instance: {
+            dbName: 'jest',
+        },
+        binary: {
+            version: '4.0.3',
+            skipMD5: true,
+        },
+        autoStart: false,
+    },
+};
